Move build status view widget
