+++ metrics/web/test/test_utils/finder_util.dart
@@ -8,7 +8,7 @@ import 'package:metrics/base/presentation/widgets/decorated_container.dart';
 import 'package:metrics/base/presentation/widgets/info_dialog.dart';
 import 'package:metrics/base/presentation/widgets/rive_animation.dart';
 import 'package:metrics/base/presentation/widgets/svg_image.dart';
-import 'package:metrics/common/presentation/build_status/widgets/build_status_view.dart';
+import 'package:metrics/common/presentation/widgets/build_status_view.dart';
 
 /// A utility class needed to find widgets in the widget tree under tests.
 class FinderUtil {
